separate out utility functions into different modules
