+++ ngwmn/fetch_utils.py
@@ -1,28 +1,12 @@
 """
-Utility functions
+Utility functions for fetching data
 
 """
 from urllib.parse import urljoin
 
-from lxml import etree
 import requests as r
 
-
-def parse_xml(xml_string):
-    """
-    Parse an XML string into an lxml object.
-    If the XML is invalid, None is returned.
-
-    :param str xml_string: a string of XML
-    :return: lxml object representing the XML content
-    :rtype: etree._Element or None
-
-    """
-    try:
-        parsed = etree.fromstring(xml_string)
-    except etree.XMLSyntaxError:
-        parsed = None
-    return parsed
+from .xml_utils import parse_xml
 
 
 def get_well_lithography(service_root, agency_cd, location_id):
