+++ ngwmn/tests/test_fetch_utils.py
@@ -1,28 +1,12 @@
 """
-Unit tests for utility functions
+Unit tests for data fetch utility functions
 """
 
 from unittest import TestCase, mock
 
-from lxml import etree
 import requests as r
 
-from ..utils import get_well_lithography, parse_xml
-
-
-class TestParseXml(TestCase):
-
-    def setUp(self):
-        self.good_xml = '<a><b>SOME TEXT</b></a>'
-        self.bad_xml = 'Ich habe eine Katze.'
-
-    def test_good_xml(self):
-        result = parse_xml(self.good_xml)
-        self.assertIsInstance(result, etree._Element)
-
-    def test_bad_xml(self):
-        result = parse_xml(self.bad_xml)
-        self.assertIsNone(result)
+from ..fetch_utils import get_well_lithography
 
 
 class TestGetWellLithography(TestCase):
